[Misc] Remove gcc `-Wsuggest-attribute=const` and `-Wsuggest-attribute=pure` warnings
These are very noisy, prone to false positives especially around assertions
and have cause me lots of troubles in the past.

They should also not be needed since most of our functions are constexpr
and or templated and thus the compiler should have a very easy time figuring
out if a function can be optimized using common subexpression elimination.

Also this is only ever useful when not using link time optimizations,
which should be discouraged.